refactor: bundle auth api in to clean architecture design.
