+++ api/data/user/user.ts
@@ -1,6 +1,6 @@
 import { NextApiResponse, NextApiRequest } from "next";
 import { AuthMiddleware } from "../../middlewares/auth.middleware";
-import { userRepository } from "../../repositories/user.repo";
+import { userRepository } from "../../repositories/user";
 import { createResponse } from "../../types/response";
 export const handleUserAPI = async (
   req: NextApiRequest,
@@ -14,8 +14,8 @@ export const handleUserAPI = async (
         .send(createResponse(false, `Unauthorized: ${error}`, null));
     }
     if (req.method === "GET") {
-     const user =  await userRepository.findOne(payload.uid);
-     return res.status(200).send(createResponse(true, "", user))
+      const user = await userRepository.findOne({ uid: payload.uid });
+      return res.status(200).send(createResponse(true, "", user));
     }
   } catch (error) {
     res.status(500).send(createResponse(false, "Internal Server Error", null));
