+++ api/usecases/index.ts
@@ -1,2 +1,2 @@
 export { applicationUsecase, ApplicationUsecase } from "./application";
-
+export { userUsecase, UserUsecase } from "./user";
