--- pages/api/auth/validate.ts
@@ -1,1 +1,0 @@
-export { handleValidate as default } from "../../../api/auth/validate";
